+++ app/src/main/java/com/aeternam/notesappmvi/feature_note/presentation/add_edit_note_screen/AddEditNoteStateHolder.kt
@@ -3,7 +3,7 @@ package com.aeternam.notesappmvi.feature_note.presentation.add_edit_note_screen
 data class AddEditNoteStateHolder(
     val title: NoteTextFieldState = NoteTextFieldState(hint = "Enter title"),
     val content: NoteTextFieldState = NoteTextFieldState(hint = "Enter some content"),
-    val color: Int = -1
+    val color: Long = -1
 )
 
 data class NoteTextFieldState(
