+++ app/src/main/java/com/aeternam/notesappmvi/feature_note/presentation/destinations/AddEditNoteDestination.kt
@@ -3,4 +3,4 @@ package com.aeternam.notesappmvi.feature_note.presentation.destinations
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class AddEditNoteDestination(val noteId: Int = -1, val noteColor: Int = -1)
+data class AddEditNoteDestination(val noteId: Int = -1, val noteColor: Long = -1)
