+++ app/src/main/java/com/aeternam/notesappmvi/feature_note/presentation/notes_screen/screen/components/NoteItem.kt
@@ -51,7 +51,7 @@ fun NoteItem(
                     cornerRadius = CornerRadius(cornerRadius.toPx())
                 )
                 drawRoundRect(
-                    color = Color(ColorUtils.blendARGB(note.color, 0x000000, 0.2f)),
+                    color = Color(ColorUtils.blendARGB(note.color.toInt(), 0x000000, 0.2f)),
                     topLeft = Offset(size.width - cutCornerSize.toPx(), -100f),
                     size = Size(cutCornerSize.toPx() + 100f, cutCornerSize.toPx() + 100f),
                     cornerRadius = CornerRadius(cornerRadius.toPx())
